+++ pkg/api/auth.go
@@ -5,9 +5,8 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/sirupsen/logrus"
-
 	"github.com/pyroscope-io/pyroscope/pkg/model"
+	"github.com/pyroscope-io/pyroscope/pkg/server/httputils"
 )
 
 const JWTCookieName = "pyroscopeJWT"
@@ -21,15 +20,13 @@ type AuthService interface {
 }
 
 // AuthMiddleware authenticates requests.
-func AuthMiddleware(log logrus.FieldLogger, loginRedirect http.HandlerFunc, authService AuthService) func(next http.Handler) http.Handler {
+func AuthMiddleware(loginRedirect http.HandlerFunc, authService AuthService, h httputils.Utils) func(next http.Handler) http.Handler {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			logger := Logger(r, log)
-
 			if token, ok := extractTokenFromAuthHeader(r.Header.Get("Authorization")); ok {
 				k, err := authService.APIKeyFromToken(r.Context(), token)
 				if err != nil {
-					Error(w, logger, model.AuthenticationError{Err: err})
+					h.HandleError(r, w, model.AuthenticationError{Err: err})
 					return
 				}
 				next.ServeHTTP(w, r.WithContext(model.WithAPIKey(r.Context(), k)))
@@ -40,24 +37,24 @@ func AuthMiddleware(log logrus.FieldLogger, loginRedirect http.HandlerFunc, auth
 				var u model.User
 				if u, err = authService.UserFromJWTToken(r.Context(), c.Value); err != nil {
 					if loginRedirect != nil {
-						logger.WithError(err).Debug("failed to authenticate jwt cookie")
+						h.Logger(r).WithError(err).Debug("failed to authenticate jwt cookie")
 						loginRedirect(w, r)
 						return
 					}
-					Error(w, logger, model.AuthenticationError{Err: err})
+					h.HandleError(r, w, model.AuthenticationError{Err: err})
 					return
 				}
 				next.ServeHTTP(w, r.WithContext(model.WithUser(r.Context(), u)))
 				return
 			}
 
-			logger.Debug("unauthenticated request")
+			h.Logger(r).Debug("unauthenticated request")
 			if loginRedirect != nil {
 				loginRedirect(w, r)
 				return
 			}
 
-			Error(w, nil, model.ErrCredentialsInvalid)
+			h.HandleError(r, w, model.ErrCredentialsInvalid)
 		})
 	}
 }