+++ pkg/server/export.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/sirupsen/logrus"
+	"github.com/pyroscope-io/pyroscope/pkg/server/httputils"
 )
 
 type Payload struct {
@@ -15,19 +15,19 @@ type Payload struct {
 }
 
 func (ctrl *Controller) exportHandler() http.HandlerFunc {
-	return NewExportHandler(ctrl.log)
+	return NewExportHandler(ctrl.httpUtils)
 }
 
-func NewExportHandler(log *logrus.Logger) http.HandlerFunc {
+func NewExportHandler(httpUtils httputils.Utils) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		resp, err := http.Post("https://flamegraph.com/api/upload/v1", "application/json", r.Body)
 		if err != nil {
-			WriteError(log, w, 500, err, fmt.Sprintf("could not upload profile: %v", err))
+			httpUtils.WriteError(r, w, 500, err, fmt.Sprintf("could not upload profile: %v", err))
 			return
 		}
 		defer resp.Body.Close()
 		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(200)
+		w.WriteHeader(http.StatusOK)
 		io.Copy(w, resp.Body)
 	}
 }